Rename Change\Attendee to ChangeAttendee
