+++ app/Models/Change.php
@@ -32,7 +32,7 @@ class Change extends Model
 
     public function attendees(): HasMany
     {
-        return $this->hasMany(Change\Attendee::class);
+        return $this->hasMany(ChangeAttendee::class);
     }
 
     public function comments(): HasMany
